Added --annotate flag
Arrow angle is determined by the slope of the line it is pointing to. It
will either be pi/3 or 2*pi/3. Unfortunately, I couldn't figure out how
to get the arrow to point perpendicular to the line, so I am pushing
this current version and will work on perpendicularity later.